backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
sqlalchemy==2.0.30
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
# Face recognition — OpenCV YuNet detector + SFace recognizer (no TensorFlow)
opencv-python-headless>=4.5.4
numpy>=1.24.0
# Depth estimation — MiDaS-small for anti-spoofing
onnxruntime>=1.15.0
Pillow>=10.0.0
aiofiles==23.2.1
python-dotenv==1.0.1
pydantic-settings==2.3.0
httpx==0.27.0
